+++ src/GatewaySettings.php
@@ -1,15 +1,14 @@
 <?php
 
 /**
- * Title: Gateway config
+ * Title: Gateway settings
  * Description:
  * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
  * @author Remco Tolsma
- * @version 1.0.0
+ * @version 1.2.4
+ * @since 1.2.4
  */
-abstract class Pronamic_WP_Pay_GatewayConfig {
-	public $id;
+abstract class Pronamic_WP_Pay_GatewaySettings {
 
-	public $mode;
 }